Move the registration API to a separate package
